+++ smooth.h
@@ -3,6 +3,6 @@
 
 #define filterSamples   7              // filterSamples should  be an odd number, no smaller than 3
 
-int digitalSmooth(int rawIn, int *sensSmoothArray);
+uint16_t digitalSmooth(uint16_t rawIn, uint16_t *sensSmoothArray);
 
 #endif
